requirements.txt
# Web Scraper Inteligente - Dependencias Python

# Framework Web
Flask==2.3.3
Flask-CORS==4.0.0

# Scraping y Web
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# Base de Datos
sqlalchemy==2.0.21
mysql-connector-python==8.2.0

# Manipulación de Datos
pandas==2.1.3
openpyxl==3.1.2

# Utilidades
python-dateutil==2.8.2
urllib3==2.0.4
lxml==4.9.3

# Logging y Configuración
python-dotenv==1.0.0

# Desarrollo (opcional)
pytest==7.4.2
black==23.7.0
flake8==6.0.0